test_requirements.txt
# not required by setup.py..
django-ckeditor

tox
coverage
django-coverage
flake8
factory_boy
mock
selenium